+++ internal/df/movie_pace.go
@@ -16,7 +16,7 @@ func MovieFrameHoldMS(s *MovieSegment, index int) float64 {
 	if index < 0 || index >= len(s.Frames) {
 		return NativeMovieFrameMS
 	}
-	return float64(max(s.Frames[index].HoldTicks, s.MinHoldTicks)) * MovieTickMS
+	return float64(float64(max(s.Frames[index].HoldTicks, s.MinHoldTicks)) * MovieTickMS)
 }
 func MovieHasRegions(s *MovieSegment) bool {
 	for _, f := range s.Frames {
@@ -38,7 +38,7 @@ func MovieSegmentInterval(s *MovieSegment, frameCount int, audioSec float64, seg
 	interval := 0.0
 	switch {
 	case !regions && s.AudioLoops && audioSec > 0:
-		interval = math.Max(NativeMovieFrameMS, audioSec*1000/float64(frameCount))
+		interval = math.Max(NativeMovieFrameMS, float64(audioSec*1000)/float64(frameCount))
 	case regions:
 		interval = NativeMovieFrameMS
 		if len(s.AudioChunks) > 0 || len(s.Sounds) > 0 {
@@ -57,7 +57,7 @@ func MovieSegmentOnScreenMS(s *MovieSegment) float64 {
 	for i, f := range s.Frames {
 		if loc, ok := s.Sounds[strings.ToLower(f.Sound)]; f.Sound != "" && ok {
 			if a, err := DecodeAudio(s.File.Data(loc), s.File.Order); err == nil {
-				soundEnds = math.Max(soundEnds, t+float64(len(a.Samples))/float64(a.SampleRate)*1000)
+				soundEnds = math.Max(soundEnds, t+float64(float64(len(a.Samples))/float64(a.SampleRate)*1000))
 			}
 		}
 		t += MovieFrameHoldMS(s, i)
@@ -160,10 +160,10 @@ func MovieSoundtrackFor(s *MovieSegment, a *MovieSegmentAudio, interval float64,
 	}
 	predicted := 0.0
 	if interval > 0 {
-		predicted = interval * float64(frameCount) / 1000
+		predicted = float64(interval*float64(frameCount)) / 1000
 	}
 	runtime := math.Max(a.AudioSec, math.Max(predicted, onScreenMS/1000))
-	cap := max(1, int(math.Ceil(runtime*1.1*float64(a.Rate))))
+	cap := max(1, int(math.Ceil(float64(float64(runtime*1.1)*float64(a.Rate)))))
 	samples := concatMovieAudio(a.Resampled, cap)
 	return &MovieSoundtrack{Audio: Audio{SampleRate: a.Rate, Samples: samples}, Loop: s.AudioLoops && len(samples) < cap}
 }
